Co-locate lenient requirement parsing
